backend/requirements.txt
asgiref==3.7.2  
Django==4.1.11 
django-cors-headers==4.2.0  
djangorestframework==3.14.0 
pytz==2023.3 
sqlparse==0.2.4  